fix: repair the clippy gate broken by the unclip commit
The previous commit shipped two redundant i64 casts in offset_point, so
cargo clippy --workspace --all-targets --all-features -D warnings failed on
main. My command pipeline printed the clippy output but did not gate the commit
on its exit code, so the failure was visible and I committed anyway.

This removes the casts and restores the gate. Verified: fmt --check clean,
clippy -D warnings exits 0, and the full suite passes at 138 library tests with
2 ignored, plus 3 contract, 6 fixture-integrity and 3 foundation.

No behaviour changes.

+++ src/unclip.rs
@@ -195,8 +195,8 @@ fn unit_normal(from: (i64, i64), to: (i64, i64)) -> (f64, f64) {
 /// Offsets one source vertex along a normal and rounds like Clipper.
 fn offset_point(source: (i64, i64), normal: (f64, f64), distance: f64) -> (i64, i64) {
     (
-        clipper_round(source.0 as f64 + normal.0 * distance) as i64,
-        clipper_round(source.1 as f64 + normal.1 * distance) as i64,
+        clipper_round(source.0 as f64 + normal.0 * distance),
+        clipper_round(source.1 as f64 + normal.1 * distance),
     )
 }
